Moved most function definitions to lib.py
